move pyspark test client fixture into conftest
